requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.11.7
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
Deprecated==1.2.13
dill==0.3.5.1
Django==3.2.14
django-celery-results==2.4.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
packaging==21.3
pep8==1.7.1
platformdirs==2.5.2
prompt-toolkit==3.0.30
pylint==2.14.5
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
typed-ast==1.5.4
typing-extensions==4.3.0
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
